fgets init commit

+++ dynamically_linked_implementation/libsabo.c
@@ -106,7 +106,8 @@ char* gets(char *src) {
  * Stop when a newline has been read, or the count runs out.
  * Return first argument, or NULL if no characters were read.
  */
-char *fgets(char *buf, size_t n, FILE *fp) {   
+char *fgets(char *buf, int n, FILE *fp)
+{   
     size_t max = isWritable(buf);
     if ( max == -1 ) {
         return NULL;
